refactor: change imports to alias, implement barrel exports
