+++ src/components/index.ts
@@ -1,0 +1,7 @@
+export { ErrorBoundary } from './error-boundary/error-boundary';
+export { FallbackUIPage } from './error-boundary/fallback-ui-page';
+export { ErrorButton } from './header/error-button';
+export { Header } from './header/header';
+export { Card } from './main/card';
+export { CardsWrapper } from './main/cards-wrapper';
+export { SearchForm } from './search-form/search-form';
