+++ src/pages/main-page.state-management.test.tsx
@@ -8,9 +8,10 @@ import { delay, http, HttpResponse } from 'msw';
 import { setupServer, SetupServerApi } from 'msw/node';
 import { expect } from 'vitest';
 
-import { apiService } from '../services/api-service';
-import { searchQueryStorage } from '../services/local-storage';
-import { testCharactersSearchArray as testCharactersSearchArray } from '../test-utils/test-data';
+import { apiService } from '@/services/api-service';
+import { searchQueryStorage } from '@/services/local-storage';
+import { testCharactersSearchArray } from '@/test-utils/test-data';
+
 import MainPage from './main-page';
 
 let server: SetupServerApi;
