+++ src/utils/guards.ts
@@ -1,7 +1,7 @@
 import {
   CharacterSearchResponse,
   SearchResponse,
-} from '../model/types-star-wars';
+} from '@/model/types-star-wars';
 
 const validateCharacterSearchResponse = (
   object: unknown,
