+++ tsconfig.app.json
@@ -10,6 +10,12 @@
 
     "types": ["vitest/globals", "@testing-library/jest-dom"],
 
+    /* Alias support */
+    "baseUrl": "./",
+    "paths": {
+      "@/*": ["src/*"]
+    },
+
     /* Bundler mode */
     "moduleResolution": "bundler",
     "allowImportingTsExtensions": true,
